requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
anthropic>=0.42.0
PyMuPDF>=1.25.0
Pillow>=11.0.0
python-dotenv>=1.0.0
pydantic-settings>=2.7.0
python-multipart>=0.0.20
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
